scripts/requirements.txt
# 数据迁移脚本依赖
psycopg2-binary>=2.9.9
pymysql>=1.1.0
